requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
python-docx==0.8.11 